requirements.txt
feedparser==6.0.11
requests==2.31.0
openai==1.30.1
python-dotenv==1.0.1
pyyaml==6.0.1
beautifulsoup4==4.12.3
lxml==5.2.2
schedule==1.2.1
rich==13.7.1
